+++ src/files/lecture.ts
